feat: created JS modules
