+++ js/address.js
@@ -1,0 +1,1 @@
+// Реализация выбора адреса на карте
