+++ js/form.js
@@ -1,0 +1,1 @@
+// Реализация формы: заполнение, условия заполнения
