+++ js/map.js
@@ -1,0 +1,1 @@
+// Реализация активного и неактивного состояний карты
